Fix #98: Some files won't open.
In the last commit, the regex that describes the `version` property was broken. Spurious backslash characters were introduced, and the optional whitespace before the colon character (which _is_ permitted by the JSON spec) was removed.

Also:
- Even though `ErrorListService` as a whole is not unit testable (I filed #100), I extracted a small static method that I could test. This is the Sprout Method (59) pattern from Feathers, "Working Effectively With Legacy Code".

Adding the unit tests revealed another bug: the extension would not have opened a SARIF file that was fewer than 200 characters long.