More efficient tagging and caret watching for VS's editor (#191)
This address issues #185 and #183.

We've had reports that the performance of Visual Studio's editor was sluggish after selecting an SARIF result from the error list and having the SARIF pane (tool window) visible.

I would go beyond that and say it wasn't usable at all :) You couldn't actually get the caret in the editor inside any highlighted text that the SARIF extension highlighted.

All of this was due to an event loop that didn't end up in a stack overflow, but a never ending "go here, go there, select this, select that" events being queued to the VS UI thread to handle. Result, super super slow VS performance.

I worked with a Visual Studio developer who specifically works on the editor to go through this issue and make sure that my changes where appropriate. After this change, the editor is usable and it solve this immediate issue.

Notable changes:

Using "persistent span factories" to create the tags/highlights so that if the user closes and re-opens a file with issues after making edits the tags/highlights remain intact.
Instead of each tag listening to a text view's caret change notification, the tagger responsible for the tags/highlights has one listener per view. Thereby increasing the performance of moving around the editor. The event firing also does not continue repeatedly if the user moves the caret within the same tag/highlight.
Highlights are properly removed when an issue is "cleared" either via the interop APIs or the error list UI.