SARIF location tagger does not need to throw on no file-name (#195)
This addresses issue #193

VS allows a text buffer to be created without file names at any time.

Our tagger requires a file name to be able to associate a SARIF result location with that file name.

So, if we don't have a file name, then don't return a tagger. Visual Studio's tagger aggregator expects and correct handles a return value of null.

This also removes some unnecessary code for retrieving a tagger from an existing text buffer (VS already has APIs for that on the text buffer).

Beyond that, the "result text marker" was caching certain Visual Studio interfaces (such as IWpfTextView, etc.) that cannot be cached and re-used as when a view is closed (users closes a document) they are not longer valid. So the "result text markers" no longer cache any Visual Studio interfaces.

Furthermore, the code flow steps were not being properly highlighted because the code was re-mapping the original region from the SARIF log over and over again via a line like "this.Region = Remap(this.Region)" (oops) so now the remapping is kept separate from the original region.