Offload log parsing from UI thread (#202)
This change partially addresses #160 by parsing SARIF on background tasks as much as possible.

The file reading and conversion (if necessary) are now being done on a background task. This alone makes the UI much more responsive while the log files are loading.

While the log file are being parsed, status is presented in Visual Studio's task center.

There is remaining work that can be done. For example, the data models can indeed be constructed on any thread and modified on any thread IF they are made thread safe and properly send change notifications on the UI thread for WPF/XAML. Also the "ruin data cache" needs to be made thread safe as well.

The error list implementation is already thread safe.

How verified:
Was able to use the editor to edit source code while a code analysis build and log parsing was in progress.
All unit-tests also pass.