Address caret tag entry (and SARIF explorer window selection) being off by one column (#219)
This fix addresses issue #218 where the tags in the documents representing SARIF locations is a bit to eager (off by one column).

The issue was rather simple. A caret's insertion point is a zero-length span, not a span of one character. The code was incorrectly adding one to the caret position to look for intersection with tags rather than just the "insertion point".