Send an initial update of the caret position for newly created text views (#221)
This address issue #220 .

When a new text view is created, and a tagger is requested, the extension creates a new text-view listener for it. Consumers of that listener (such as the SARIF explorer window) use the events from the text-view listener to properly show selected items in their views.

During initial creation of the text view (and its corresponding listener), an initial set of events was not being sent for the caret position in the text view. Due to this, the SARIF explorer XAML views could not properly update their initial selection.

The fix was to send an initial events for the starting caret location for the text view.

While testing the fix, I also realized that when you clear the SARIF errors from the list, or do anything else that results in a SARIF result not being valid (such as closing a SARIF log file that was directly opened, etc.), that the SARIF explorer continued to show content for a SARIF result that has been disposed. That's not so good. This was also addressed.