Ensure telemetry provider is initialized regardless of package initialization status (#226)
This change address issue #216 and ensure that telemetry is properly sent from the SARIF extension.

The implementation uses MEF to export a tiny telemetry provider type\service that is constructed on demand when the first telemetry event is fired. The reason MEF is used is that it provides a way to "flush" and "dispose" the telemetry client on Visual Studio shutdown (confirmed with debugger and Visual Studio developers).

The implementation provides an easy and consistent way to add new telemetry points in a way that allows for easy "auditing" of the telemetry as well as consistent naming.