Fix #236: Provide an interop API to accept a SARIF log from a stream (#259)
This turns out to require only a small change to `ErrorListService`. We introduce a "shim" API `ErrorListService.ProcessSarifLogAsync(Stream)`, which deserializes the stream into a `SarifLog` and then calls an existing private API `ErrorListService.ProcessSarifLogAsync(SarifLog)`.

The fact that there is no file name available turns out not to matter:
- The file name never appears in the UI.
- At the moment, there is no gesture to re-read the file, or to re-persist it, so the lack of a backing file is moot.
- The extension does use the log file path, when prompting the user to resolve a source file path, as the "initial directory" in the File Open dialog. But if the path is null, it just starts in the current working directory.

The only other thing the extension uses the log file path for is as a key into a dictionary that associates sets of results with log files. This allows the extension to remove those results from the Error List when the user closes a log file. Without a file name, the results from a streamed in file can't be collectively removed from the error list.

But the scenario for using the stream-based APIs is background analysis, and there's no gesture today to remove (for example) all IDE warnings from the error list, or all warnings from the StyleCop or FxCop analyzer. So our proposed SARIF-based background analysis is no worse off in this regard than existing background analyses.

In fact, the stream-based API is no worse off than the file-based API in this regard. When another extension sends a SARIF file to the viewer through the file-based API, the viewer does _not_ display the file -- so again, there is no "close the file and remove its results from the Error List" gesture.

Because this turns out to "just work", we don't need to refactor, so we will close #254, "Refactors in preparation for stream-driven SARIF transfer API", as `resolved-wont-fix`.