Allow exceptions to escape from task status center so tests can see them (#287)
In the file-based API code path, the log file handling was being done in a background task in the Task Status Center. If that background task threw an exception, it was lost. The affected test would fail with an verification failure, but the stack trace did not reveal the exception that caused the failure.

In this PR, we refactor to assign the background task to a temporary variable, pass the task variable to the Task Status Center, and then await the task ourselves. Now if the task throws an exception, the viewer will rethrow it, and Apex will detect and report it.

I've tested this by intentionally introducing an exception into the error list service's log file handling and watching Apex report it.